feat: API 테스트 케이스 작성
FDORA-35
