+++ src/test/java/com/farmdora/farmdora/sale/controller/SaleControllerTest.java
@@ -1,119 +1,34 @@
 package com.farmdora.farmdora.sale.controller;
 
+import static com.farmdora.farmdora.common.response.SuccessMessage.GET_RELATED_SALES_SUCCESS;
 import static com.farmdora.farmdora.common.response.SuccessMessage.GET_SALE_DETAIL_SUCCESS;
-import static com.farmdora.farmdora.common.response.SuccessMessage.SEARCH_SALES_SUCCESS;
 import static org.hamcrest.Matchers.equalTo;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.farmdora.farmdora.ControllerTest;
 import com.farmdora.farmdora.common.exception.ResourceNotFoundException;
-import com.farmdora.farmdora.common.response.PageResponseDto;
-import com.farmdora.farmdora.order.dto.Sort;
 import com.farmdora.farmdora.sale.dto.SaleDetailDto;
 import com.farmdora.farmdora.sale.dto.SaleDetailDto.OptionDetailDto;
-import com.farmdora.farmdora.sale.dto.SaleSearchRequestDto;
-import com.farmdora.farmdora.sale.dto.SaleSearchResponseDto;
-import com.farmdora.farmdora.sale.dto.SaleStatus;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import com.farmdora.farmdora.sale.dto.SaleRelatedDto;
 import java.util.List;
-import java.util.Set;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.MediaType;
 
-class SaleControllerTest extends ControllerTest {
-
-    private SaleSearchRequestDto searchCondition;
-    private PageResponseDto<SaleSearchResponseDto> result;
-
-    @BeforeEach
-    void setUp() {
-        searchCondition  = SaleSearchRequestDto.builder()
-                .keyword("상추")
-                .sort(Sort.LATEST)
-                .filters(Set.of(SaleStatus.INSTOCK))
-                .typeBigId((short) 1)
-                .typeId((short) 2)
-                .build();
-
-        List<SaleSearchResponseDto> saleSearchResponseDtos = List.of(
-                SaleSearchResponseDto.builder()
-                        .saleId(1)
-                        .title("상추1")
-                        .price(10000)
-                        .isBlind(false)
-                        .stock(100)
-                        .orderCount(3L)
-                        .build(),
-                SaleSearchResponseDto.builder()
-                        .saleId(2)
-                        .title("상추2")
-                        .price(20000)
-                        .isBlind(false)
-                        .stock(200)
-                        .orderCount(2L)
-                        .build()
-        );
-        Page<SaleSearchResponseDto> pages = new PageImpl<>(saleSearchResponseDtos);
-        result = new PageResponseDto(pages.getContent(), pages);
-    }
-
-    @Nested
-    @DisplayName("상품 목록 조회 API 테스트")
-    class SearchSales {
-
-        @Test
-        @DisplayName("JSON으로 상품 목록 조회 API 테스트")
-        void testSearchSalesByJsonAPI() throws Exception {
-            // given
-            when(saleService.searchSales(anyInt(), any(SaleSearchRequestDto.class), any(Pageable.class))).thenReturn(result);
-
-            // when
-            // then
-            mvc.perform(post("/my/seller/sale/search")
-                            .contentType(MediaType.APPLICATION_JSON)
-                            .content(new ObjectMapper().writeValueAsString(searchCondition)))
-                    .andExpect(status().isOk())
-                    .andExpect(jsonPath("$.status", equalTo(200)))
-                    .andExpect(jsonPath("$.message", equalTo(SEARCH_SALES_SUCCESS.getMessage())));
-        }
-
-        @Test
-        @DisplayName("파라미터로 상품 목록 조회 API 테스트")
-        void testSearchSalesByParamsAPI() throws Exception {
-            // given
-            when(saleService.searchSales(anyInt(), any(SaleSearchRequestDto.class), any(Pageable.class))).thenReturn(result);
-
-            // when
-            // then
-            mvc.perform(get("/my/seller/sale/search")
-                            .param("keyword", "상추")
-                            .param("sort", "LATEST")
-                            .param("typeBigId", "1")
-                            .param("typeId", "2"))
-                    .andExpect(status().isOk())
-                    .andExpect(jsonPath("$.status", equalTo(200)))
-                    .andExpect(jsonPath("$.message", equalTo(SEARCH_SALES_SUCCESS.getMessage())));
-        }
-    }
+public class SaleControllerTest extends ControllerTest {
 
     @Nested
     @DisplayName("상품 상세 조회 API 테스트")
     class GetSaleDetail {
 
         @Test
-        @DisplayName("상품 상세 조회 API 테스트")
+        @DisplayName("상품 상세 조회 API 성공")
         void testGetSaleDetailAPI() throws Exception {
             // given
             SaleDetailDto saleDetailDto = SaleDetailDto.builder()
@@ -127,7 +42,7 @@ class SaleControllerTest extends ControllerTest {
 
             // when
             // then
-            mvc.perform(get("/my/seller/sale/{saleId}", 1)
+            mvc.perform(get("/sale/{saleId}", 1)
                             .param("userId", "1"))
                     .andExpect(status().isOk())
                     .andExpect(jsonPath("$.status", equalTo(200)))
@@ -144,7 +59,7 @@ class SaleControllerTest extends ControllerTest {
 
             // when
             // then
-            mvc.perform(get("/my/seller/sale/{saleId}", 1)
+            mvc.perform(get("/sale/{saleId}", 1)
                             .param("userId", "1"))
                     .andExpect(status().isBadRequest())
                     .andExpect(jsonPath("$.status", equalTo(400)))
@@ -152,5 +67,50 @@ class SaleControllerTest extends ControllerTest {
         }
     }
 
+    @Nested
+    @DisplayName("관련 상품 목록 조회 API 테스트")
+    class GetRelatedSales {
+        @Test
+        @DisplayName("관련 상품 목록 조회 API 성공")
+        void testGetRelatedSalesAPI() throws Exception {
+            // given
+            List<SaleRelatedDto> relatedSales = List.of(
+                    SaleRelatedDto.builder()
+                            .saleId(1)
+                            .isLike(true)
+                            .reviewCount(10L)
+                            .build(),
+                    SaleRelatedDto.builder()
+                            .saleId(1)
+                            .isLike(true)
+                            .reviewCount(10L)
+                            .build()
+            );
+            when(saleService.getRelatedSales(anyInt(), anyInt(), any(Pageable.class))).thenReturn(relatedSales);
 
+            // when
+            // then
+            mvc.perform(get("/sale/related/{saleId}", 1)
+                            .param("userId", "1"))
+                    .andExpect(status().isOk())
+                    .andExpect(jsonPath("$.status", equalTo(200)))
+                    .andExpect(jsonPath("$.message", equalTo(GET_RELATED_SALES_SUCCESS.getMessage())))
+                    .andExpect(jsonPath("$.data.size()", equalTo(2)));
+        }
+
+        @Test
+        @DisplayName("관련 상품 목록 조회시 상품이 존재하지 않을 경우 예외 발생 API 테스트")
+        void testGetRelatedSalesAPI_SaleNotFoundException() throws Exception {
+            // given
+            when(saleService.getRelatedSales(anyInt(), anyInt(), any(Pageable.class))).thenThrow(new ResourceNotFoundException("Sale", 1));
+
+            // when
+            // then
+            mvc.perform(get("/sale/related/{saleId}", 1)
+                            .param("userId", "1"))
+                    .andExpect(status().isBadRequest())
+                    .andExpect(jsonPath("$.status", equalTo(400)))
+                    .andExpect(jsonPath("$.message", equalTo("Sale 데이터가 존재하지 않습니다 : '1'")));
+        }
+    }
 }
